+++ README.md
@@ -5,7 +5,8 @@ CREATE TABLE users (
     user_id SERIAL PRIMARY KEY,
     email VARCHAR(45) NOT NULL,
     username VARCHAR(25) NOT NULL,
-    password VARCHAR(250) NOT NULL
+    password VARCHAR(250) NOT NULL,
+    history JSON
 )
 ```
