+++ src/components/Blacklist.js
@@ -1,0 +1,5 @@
+const Blacklist = () => {
+
+}
+
+export default Blacklist
